Fixed key states of WtlApplicationImpl::trigger without focus.

+++ src/good/app/wtl_player.h
@@ -304,7 +304,7 @@ public:
       if (MAX_CMD_LINE_HIST < cmdHist.size()) {
         cmdHist.erase(cmdHist.begin());
       }
-      iCmdHist = cmdHist.size();
+      iCmdHist = (int)cmdHist.size();
       lazyRunCmdLine = cmdLine;
       cmdLine = "";
       break;
